backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
torch==2.6.0
torchvision==0.21.0
diffusers==0.32.0
transformers==4.48.0
accelerate==1.3.0
safetensors==0.5.0
mediapipe==0.10.32
aiohttp>=3.11.0
redis==5.2.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2

